requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.37.0
Pillow>=8.3.0
accelerate>=0.20.0
tiktoken>=0.5.0
